# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API del sistema hospitalario Galenos Pro. Arquitectura hexagonal, persiste en SQL Server 2022.",
        "title": "Galenos Pro Appointments API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sis/consumo": {
            "get": {
                "description": "Invoca el SP webFactOrdenServicioDetaDesFinaListarIdcuenta con el id de la cuenta de atención.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SIS"
                ],
                "summary": "Lista el detalle de consumo de una atención",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id de la cuenta de atención",
                        "name": "idCuentaAtencion",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de consumos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parámetro inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al listar el consumo",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sis/diagnosticos": {
            "get": {
                "description": "Invoca el SP webAtencionesDiagnosticosIdCuentaAtencion con el id de la cuenta de atención. Devuelve los diagnósticos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SIS"
                ],
                "summary": "Consulta los diagnósticos de una atención",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id de la cuenta de atención",
                        "name": "idCuentaAtencion",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de diagnósticos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al consultar los diagnósticos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sis/fua": {
            "post": {
                "description": "Invoca el SP webSisFuaAtencionForzarGuardado para forzar la generación/guardado del FUA.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SIS"
                ],
                "summary": "Fuerza el guardado del FUA de una cuenta de atención",
                "parameters": [
                    {
                        "description": "Id de la cuenta de atención",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpadapter.sisFuaForzarGuardadoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FUA guardado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Cuerpo inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al guardar el FUA",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sis/fua/agregar": {
            "post": {
                "description": "Invoca el SP usp_go_webFUAgregar para registrar el FUA. Devuelve el @Respuesta del procedimiento.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SIS"
                ],
                "summary": "Agrega el FUA de una cuenta de atención",
                "parameters": [
                    {
                        "description": "Datos para agregar el FUA",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpadapter.sisFuaAgregarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta del SP",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Cuerpo inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al agregar el FUA",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sis/fua/imprimir": {
            "get": {
                "description": "Invoca el SP webFuaImprimirIdCuentaAtencion con el id de la cuenta de atención. Devuelve los datos del FUA.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SIS"
                ],
                "summary": "Consulta los datos del FUA para imprimir",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id de la cuenta de atención",
                        "name": "idCuentaAtencion",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del FUA",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parámetro inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al consultar el FUA",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/sis/medicamentos": {
            "get": {
                "description": "Invoca el SP webMedicamentosListarIdCuentaAtencion con el id de la cuenta de atención.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SIS"
                ],
                "summary": "Lista los medicamentos de una atención",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id de la cuenta de atención",
                        "name": "idCuentaAtencion",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de medicamentos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Parámetro inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error al listar los medicamentos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpadapter.apiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/httpadapter.apiError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpadapter.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpadapter.apiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/httpadapter.apiError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "httpadapter.sisFuaAgregarRequest": {
            "type": "object",
            "required": [
                "idCuentaAtencion",
                "idEmpleado"
            ],
            "properties": {
                "idCuentaAtencion": {
                    "type": "integer"
                },
                "idEmpleado": {
                    "type": "integer"
                },
                "nombrePc": {
                    "type": "string"
                }
            }
        },
        "httpadapter.sisFuaForzarGuardadoRequest": {
            "type": "object",
            "required": [
                "idCuentaAtencion"
            ],
            "properties": {
                "idCuentaAtencion": {
                    "type": "integer"
                }
            }
        },
...
    }
}